extracted studentlist functions
